requirements.txt

matplotlib
numpy
numba
galois
pandas
scipy
vprof
rustworkx
Cython
jsonlines
python-dateutil
tqdm
